requirements.txt
APScheduler==3.0.5
Django==1.7
Eve==0.6.1
Events==0.2.1
Pillow==2.8.1
PyMySQL==0.7.1
pymongo==3.6.0
pytz==2016.3
redis==2.10.3
requests==2.7.0
six==1.10.0
tzlocal==1.2.2
Sphinx==1.3.1
sphinx-py3doc-enhanced-theme==2.2.4
sphinx-rtd-theme==0.1.9
bson==0.5.1
pexpect==4.3.1
